// src/views/menu/charities/add-charity/data/types.ts
export interface Charity {
  accountsDate: string;
  address: string;
  beneficiary: string;
  category: string;
  charityNumber: string;
  charityURL: string;
  collectedDonation: string;
  contactEmail: string;
  contactName: string;
  contactNumber: string;
  createdBy: string;
  country: string;
  coverPicture: "";
  displayPicture: "";
  documents: "";
  facebook: string;
  feed: string;
  id: string;
  instagram: string;
  isFeatured: boolean;
  isVerified: boolean;
  lastChecked: string;
  likes: number;
  longDescription: string;
  operation: string;
  private: boolean;
  registrationDate: string;
  removeDate: string;
  shortDescription: string;
  startDate: string;
  status: string;
  subtitle: string;
  title: string;
  totalIncome: string;
  totalSpent: string;
  twitter: string;
  URL: string;
  user: string;
  views: number;
  website: string;
  youtube: string;
}